requirements.txt
boto==2.48.0
boto3==1.5.22
botocore==1.8.36
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
dj-database-url==0.4.2
Django==1.11.10
django-crispy-forms==1.7.0
django-extensions==1.9.9
django-filter==1.1.0
django-storages==1.6.5
djangorestframework==3.7.7
docutils==0.14
gunicorn==19.7.1
idna==2.6
Jinja2==2.10
jmespath==0.9.3
lxml==3.8.0
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
olefile==0.45.1
Pillow==5.0.0
psycopg2==2.7.3.2
PyJWT==1.5.3
PySocks==1.6.8
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
twilio==6.10.2
typing==3.6.4
urllib3==1.22
